+++ tests/functional/test_issue_reproductions.py
@@ -303,3 +303,29 @@ class TestIssue121(unittest.TestCase):
 
         self.assertFalse(self.repo.git.settings('list', '--local'))
         self.assertFalse(self.repo.git.settings('list', '--file', self.dirpath + '/.git/config'))
+
+
+class TestIssue122(unittest.TestCase):
+    """Settings list does not gracefully handle unknown files"""
+
+    def setUp(self):
+        self.dirpath = tempfile.mkdtemp()
+        os.chdir(self.dirpath)
+
+        # initialize repository
+        self.repo = git.Repo.init(self.dirpath)
+
+    def tearDown(self):
+        shutil.rmtree(self.dirpath)
+
+    def test(self):
+        """Issue 122: settings list should gracefully print unknown file errors"""
+
+        associate_proc = subprocess.Popen(
+            'git settings list --file unknown_file'.split(),
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE
+        )
+        stdout, stderr = associate_proc.communicate()
+        self.assertFalse(stdout)
+        self.assertEqual(stderr.strip(), "error: no such file 'unknown_file'")