+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file. This projec
 - `changes unassociate --dry-run` printing a message during a no-op [#108](https://github.com/Brickstertwo/git-commands/issues/108)
 - `changes associate --upstream` failing with local upstreams [#112](https://github.com/Brickstertwo/git-commands/issues/112)
 - `settings list` failing with empty config files [#121](https://github.com/Brickstertwo/git-commands/issues/121)
+- `settings list` not properly handling unknown files [#122](https://github.com/Brickstertwo/git-commands/issues/122)
 
 ## [v0.6.1][] - 2017-01-17
 ### Fixed